added set_body_sent

+++ src/MRPP/util/mrpp_state.c
@@ -202,8 +202,8 @@ static void update_bodies(MRPP_STATE *state, uint8_t collectionId){
     
 }
 
-int8_t mrpp_state_is_body_ready(MRPP_STATE *state){
-    int8_t bodyIndex=-1;
+int16_t mrpp_state_is_body_ready(MRPP_STATE *state){
+    int16_t bodyIndex=-1;
     for (uint8_t i = 0; i < state->nBodies; i++)
     {
         if(state->bodies[i]==READY){
@@ -214,8 +214,8 @@ int8_t mrpp_state_is_body_ready(MRPP_STATE *state){
     return bodyIndex;
 }
 
-bool mrpp_state_get_ready_body(MRPP_STATE *state,uint8_t bodyIndex, uint8_t *subId, uint8_t *lastSubId, uint16_t *begin, uint8_t*length){
-    int8_t res=mrpp_state_is_body_ready(state);
+bool mrpp_state_get_ready_body(MRPP_STATE *state, int16_t bodyIndex, uint8_t *subId, uint8_t *lastSubId, uint16_t *begin, uint8_t*length){
+    int16_t res=mrpp_state_is_body_ready(state);
     if(res==-1 || res!=bodyIndex ) return false;
 
     uint8_t readyIndex;
@@ -239,7 +239,7 @@ bool mrpp_state_get_ready_body(MRPP_STATE *state,uint8_t bodyIndex, uint8_t *sub
     return true; 
 }
 
-void mrpp_state_set_body_sent(MRPP_STATE *state, uint8_t bodyIndex){
+void mrpp_state_set_body_sent(MRPP_STATE *state, int16_t bodyIndex){
     if(state->nBodies<=bodyIndex) return;
 
     state->bodies[bodyIndex]=SENT;
